+++ data-dashboard-backend/src/test/java/org/radarbase/datadashboard/api/util/MockAsyncCoroutineService.kt
@@ -1,4 +1,22 @@
-package org.radarbase.datadashboard.api.mock
+/*
+ *
+ *  *  Copyright 2026 The Hyve
+ *  *
+ *  *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  *  you may not use this file except in compliance with the License.
+ *  *  You may obtain a copy of the License at
+ *  *
+ *  *    http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *  Unless required by applicable law or agreed to in writing, software
+ *  *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *  See the License for the specific language governing permissions and
+ *  *  limitations under the License.
+ *
+ */
+
+package org.radarbase.datadashboard.api.util
 
 import jakarta.ws.rs.container.AsyncResponse
 import kotlinx.coroutines.CancellableContinuation
@@ -25,7 +43,8 @@ class MockAsyncCoroutineService : AsyncCoroutineService {
 
     override suspend fun <T> runInRequestScope(block: () -> T): T = block()
 
-    override suspend fun <T> suspendInRequestScope(block: (CancellableContinuation<T>) -> Unit): T = suspendCancellableCoroutine(block)
+    override suspend fun <T> suspendInRequestScope(block: (CancellableContinuation<T>) -> Unit): T =
+        suspendCancellableCoroutine(block)
     override suspend fun <T> withContext(name: String, block: suspend () -> T): T {
         TODO("Not yet implemented")
     }
